requirements.txt
PySide6~=6.5.2
pyserial~=3.5
opencv-python~=4.8.0.76
Unidecode~=1.3.6